feat: seccions

+++ app/src/views/prestamos/ListaView.vue
@@ -12,7 +12,7 @@
           <input type="date" class="rounded py-0.5 " v-model="startDate">
           <span class="text-gray-800 dark:text-white ">Fin</span>
           <input type="date" class="rounded py-0.5 " v-model="endDate">
-          <button class="bg-blue-500 px-4 rounded text-white" @click="searchByDate">Buscar</button>
+          <button class="bg-blue-500 px-4 rounded text-white py-1" @click="searchByDate">Buscar</button>
         </div>
       </div>
     </div>
@@ -46,7 +46,8 @@
         <tbody v-if="prestamos">
           <tr v-for="prestamo in prestamos" :key="prestamo.id" class="libro">
             <td class="px-3 py-1 text-sm font-medium text-gray-500 whitespace-nowrap">
-              <h2 class="font-medium text-gray-800 dark:text-white ">{{ prestamo.persona.nombres }} {{ prestamo.persona.apellidos }}</h2>
+              <h2 class="font-medium text-gray-800 dark:text-white ">{{ prestamo.persona.nombres }} {{
+                prestamo.persona.apellidos }}</h2>
               <p class="text-sm font-normal text-gray-600 dark:text-gray-400 -mt-1">{{ prestamo.persona.tipo.descripcion
                 }}</p>
             </td>
@@ -84,7 +85,7 @@
                   class="text-gray-500 transition-colors duration-200 dark:hover:text-blue-500 dark:text-gray-300 hover:text-blue-500 focus:outline-none">
                   <v-icon name="md-preview-outlined" class="w-7 h-7 p-1" title="Detalles"></v-icon>
                 </button>
-                <button
+                <button @click="ItemById(prestamo.id)"
                   class="text-gray-700 transition-colors duration-200 dark:hover:text-yellow-500 dark:text-gray-300 hover:text-yellow-500 focus:outline-none">
                   <v-icon name="la-edit-solid" class="w-7 h-7 p-1" title="Editar"></v-icon>
                 </button>
@@ -191,7 +192,7 @@
           </div>
           <div class="flex">
             <span class="font-semibold mx-2 "> Monto:</span>
-            <div>{{ prestamosId.multa ? prestamo.multa.monto : '0.00'}}</div>
+            <div>{{ prestamosId.multa ? prestamo.multa.monto : '0.00' }}</div>
           </div>
           <div>
             <span class="font-semibold me-2">Libros:</span>
@@ -228,9 +229,78 @@
     </div>
   </div>
 
+  <div v-if="openModalEdit" class=" absolute top-0 left-0  w-screen h-screen backdrop-blur"></div>
+  <div v-if="openModalEdit"
+    class="modalUser bg-white dark:bg-gray-800 dark:border-gray-900 p-6 rounded-md shadow border">
+    <h2 class="mb-2 font-semibold dark:text-gray-100">Edicion</h2>
+    <div v-if="ItemEdit.id" class="sm:flex sm:space-x-6  text-gray-800">
+      <div class="flex flex-row gap-2 w-full">
+        <div class="flex flex-col w-full">
+          <span class="text-sm text-gray-600 font-semibold">Persona:</span>
+          <input type="text" v-model="ItemEdit.persona.nombres" class="rounded py-1 mb-2" readonly>
+          <span class="text-sm text-gray-600 font-semibold">Fecha Inicio:</span>
+          <input type="date" v-model="ItemEdit.fecha_inicio" class="rounded py-1 mb-2">
+          <span class="text-sm text-gray-600 font-semibold">Fecha Fin:</span>
+          <input type="date" v-model="ItemEdit.fecha_fin" class="rounded py-1 mb-2">
+          <span class="text-sm text-gray-600 font-semibold mt-3">Editado:</span>
+          <span class="text-gray-600 -mt-1">{{ useTime(ItemEdit.updated_at).value }}</span>
+        </div>
+        <div class="flex flex-col w-full bg-gray-50 border rounded p-2">
+          <h3>Gestionar Libros</h3>
+          <div class="w-full">
+            <div class="flex w-full">
+              <input type="text" v-model="searchLibro" placeholder="Buscar..." class="rounded-s py-1 w-full"
+                @focus="showDropdown = true" @blur="hideDropdown" @input="handleSearchLibro()" />
+              <button @click="showDropdown = false"
+                class="rounded-e border-e border-y px-2 border-gray-500 py-1 ">X</button>
+            </div>
+            <div v-if="showDropdown && libros.length > 0" class="dropdown absolute w-[358px]">
+              <ul>
+                <li v-for="libro in libros" :key="libro.id" @click="selectOption(libro)" class="dropdown-item">
+                  {{ libro.name }}
+                </li>
+              </ul>
+            </div>
+          </div>
+          <div>
+            <h4>Lista de libros</h4>
+            <div class="flex flex-wrap gap-2">
+              <div v-for="libro in ItemEdit.libros" :key="libro.id"
+                class="bg-gray-100 border-gray-500 rounded-full border px-2 py-0.5 w-max flex items-center">
+                <span>{{ libro.titulo }} </span>
+                <button @click="deleteLibro(ItemEdit.id, libro.id)"
+                  class="text-red-500 font-semibold hover:bg-gray-200 rounded-full px-1.5 py-0 ms-1">X</button>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div v-else class="w-full max-w-md mx-auto animate-pulse p-9">
+      <h1 class="h-2 bg-gray-300 rounded-lg w-52 dark:bg-gray-600"></h1>
+      <p class="w-48 h-2 mt-6 bg-gray-200 rounded-lg dark:bg-gray-700"></p>
+      <p class="w-full h-2 mt-4 bg-gray-200 rounded-lg dark:bg-gray-700"></p>
+      <p class="w-64 h-2 mt-4 bg-gray-200 rounded-lg dark:bg-gray-700"></p>
+      <p class="w-4/5 h-2 mt-4 bg-gray-200 rounded-lg dark:bg-gray-700"></p>
+    </div>
+    <div class="mt-4 flex justify-end gap-2">
+      <button @click="editItem(ItemEdit)" class="bg-blue-500 px-4 py-1 rounded-md text-white">Guardar Cambios</button>
+      <button @click="openModalEdit = false" class="bg-gray-500 px-4 py-1 rounded-md text-white">Cerrar</button>
+    </div>
+  </div>
+
   <div v-if="openModalDelete" class=" absolute top-0 left-0 w-screen h-screen backdrop-blur"></div>
   <div v-if="openModalDelete" class="modal bg-white dark:bg-gray-800 dark:border-gray-900 p-6 rounded-md shadow border">
     <h2 class="mb-2 font-semibold dark:text-gray-100">{{ mensaje_modal }}</h2>
+
+    <div v-if="mensaje_modal == 'Finalizar'" class="flex flex-col">
+      <p class="mb-2 font-light dark:text-gray-100">Agregar un multa</p>
+      <span class="text-sm text-gray-600 font-semibold">Asunto:</span>
+      <input type="text" v-model="multaAsunto" class="rounded py-1 mb-2">
+      <span class="text-sm text-gray-600 font-semibold">Monto:</span>
+      <input type="number" v-model="multaMonto" class="rounded py-1 mb-2">
+    </div>
+
     <p class="mb-2 font-light dark:text-gray-100">¿Esta seguro de realizar esta accion?</p>
     <div class="mt-4 flex justify-end gap-2">
       <button v-if="mensaje_modal == 'Finalizar'" @click="deleteItem(item_id, 3)"
@@ -246,7 +316,7 @@
 <script setup>
 import PrestamosService from "../../services/PrestamosService.js";
 import useTime from "@/composables/useTime.js"
-import { onMounted, ref } from "vue";
+import { onMounted, ref, computed } from "vue";
 import { toast } from 'vue3-toastify';
 import 'vue3-toastify/dist/index.css';
 
@@ -257,11 +327,15 @@ const modal = (id, mensaje) => {
 }
 const openModalDetalles = ref(false)
 const openModalDelete = ref(false)
+const openModalEdit = ref(false)
 const mensaje_modal = ref('')
 
 const service = new PrestamosService();
 const prestamos = ref([]);
 const prestamosId = ref([]);
+const ItemEdit = ref([]);
+const multaAsunto = ref('-');
+const multaMonto = ref(0);
 const item_id = ref('');
 const inputBuscar = ref('');
 const startDate = ref('');
@@ -278,12 +352,8 @@ const fetchPrestamos = async (buscar, page = 1) => {
   currentPage.value = response.current_page;
   lastPage.value = response.last_page;
 };
-const handleSearch = async (input) => {
-  // if (inputBuscar.value.length >= 2 || inputBuscar.value.length === 0) {
+const handleSearch = async () => {
   await fetchPrestamos(inputBuscar.value);
-  // } else {
-  //   categorias.value = [];
-  // }
 };
 const searchByDate = async () => {
   try {
@@ -305,7 +375,11 @@ onMounted(async () => {
   fetchPrestamos('', 1);
 });
 const deleteItem = async (id, estado) => {
-  state.value = await service.deleteById(id, estado);
+  if (estado == 3) {
+    state.value = await service.addMulta(multaAsunto.value, multaMonto.value, id);
+  } else {
+    state.value = await service.deleteById(id, estado);
+  }
   if (!state.value.error) {
     toast.success("Operacion realizada", {
       autoClose: 3000,
@@ -329,9 +403,88 @@ const LibroById = async (id) => {
     console.log('Error fetching book by ID:', error);
   }
 };
+const editItem = async (item) => {
+  try {
+    await service.update(item.id, item);
+    toast.success("Operacion realizada", {
+      autoClose: 3000,
+    });
+    fetchPrestamos('', currentPage.value);
+    openModalEdit.value = false;
+  } catch (error) {
+    console.log('Error updating item:', error);
+  }
+}
+// Busqeuda
+const selectedOption = ref(null);
+const showDropdown = ref(false);
+const searchLibro = ref('');
+const options = ref([]);
+const libros = computed(() => {
+  return options.value.filter(option =>
+    option.name.toLowerCase().includes(searchLibro.value.toLowerCase())
+  );
+});
+const handleSearchLibro = async () => {
+  const dataLibros = await service.fetchLibros(searchLibro.value);
+  if (dataLibros && Array.isArray(dataLibros.data)) {
+    options.value = dataLibros.data.map(item => ({
+      id: item.id,
+      name: item.titulo
+    }))
+    showDropdown.value = true
+  }
+};
+const selectOption = async (option) => {
+  await service.addLibro(ItemEdit.value.id, option.id);
+  ItemById(ItemEdit.value.id);
+  selectedOption.value = option;
+  showDropdown.value = false;
+};
+const ItemById = async (id) => {
+  try {
+    if (id) {
+      openModalEdit.value = true;
+      const response = await service.fetchById(id);
+      ItemEdit.value = response.data[0];
+    }
+  } catch (error) {
+    console.log('Error fetching book by ID:', error);
+  }
+};
+
+const deleteLibro = async (prestamo_id, libro_id) => {
+  await service.deleteDetalleLibro(prestamo_id, libro_id);
+  toast.success("Operacion realizada", {
+    autoClose: 2500,
+  });
+  ItemById(prestamo_id);
+}
+
 </script>
 
 <style scoped>
+.dropdown {
+  border: 1px solid #ccc;
+  max-height: 200px;
+  overflow-y: auto;
+  background: white;
+}
+
+.dropdown-item {
+  padding: 8px 12px;
+  cursor: pointer;
+}
+
+.dropdown-item:hover {
+  background-color: #f0f0f0;
+}
+
+.selected-option {
+  margin-top: 10px;
+  font-weight: bold;
+}
+
 .modal {
   position: fixed;
   z-index: 999;
